REFACTOR: Moved Authentication & Music Related Classes to their own Packages
