+++ app/src/main/java/com/projectx/eyemusic/MainActivity.java
@@ -24,6 +24,9 @@ import android.widget.Toast;
 import com.android.volley.RequestQueue;
 import com.android.volley.Response;
 import com.android.volley.toolbox.Volley;
+import com.projectx.eyemusic.Authentication.Authentication;
+import com.projectx.eyemusic.Music.Playlist;
+import com.projectx.eyemusic.Music.PlaylistAdapter;
 import com.projectx.eyemusic.VolleyRequests.PlaylistRequest;
 import com.projectx.eyemusic.graphics.DotGraphic;
 import com.projectx.eyemusic.graphics.GraphicOverlay;
