+++ app/src/main/res/layout/activity_tracks.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".TracksActivity">
+    tools:context=".Music.TracksActivity">
     
 
     <androidx.appcompat.widget.Toolbar
